requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
python-dotenv
sentence-transformers
faiss-cpu
pdfplumber
ruff
langchain-ollama